Optimise Entity.TransformMatrix, prefer 32 to 64 bit usages in places, start work on parsing C string to 32 bit number.

+++ ClassicalSharp/Game/Game.Properties.cs
@@ -161,7 +161,7 @@ namespace ClassicalSharp {
 		/// <summary> Whether mouse rotation on the y axis should be inverted. </summary>
 		public bool InvertMouse;
 		
-		public long Vertices;
+		public int Vertices;
 		public FrustumCulling Culling;
 		public AsyncDownloader AsyncDownloader;
 		public Matrix4 View, Projection;
